+++ tp/game/models.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from sqlalchemy import Column, Integer, String, ForeignKey, Enum, Table, Boolean, UniqueConstraint
+from sqlalchemy import Column, Integer, String, ForeignKey, Enum, Table, Boolean, UniqueConstraint, BigInteger
 from sqlalchemy.orm import relationship
 from flask import g
 from tp.base.models import Base, CommonPK
@@ -18,10 +18,10 @@ class Game(Base, CommonPK):
   #utenti che partecipano al gioco
   users = relationship("Partecipation", back_populates = "game")
   #utente vincitore
-  winner_id = Column(Integer, ForeignKey("user.id"))
+  winner_id = Column(BigInteger, ForeignKey("user.id"))
   winner = relationship("User", foreign_keys = [winner_id])
   #the id of the player that invited other players to this game at the beginning
-  creator_id = Column(Integer, ForeignKey("user.id"))
+  creator_id = Column(BigInteger, ForeignKey("user.id"))
   creator = relationship("User", foreign_keys = [creator_id])
   #il gioco è finito?
   ended = Column(Boolean, default = False)
@@ -45,13 +45,13 @@ class Round(Base, CommonPK):
   #numero del round, TODO: trovare il modo di farlo diventare autoincrement rispetto a diversi parametri
   number = Column(Integer, nullable = False)
   #match di riferimento
-  game_id = Column(Integer, ForeignKey("game.id"), nullable = False)
+  game_id = Column(BigInteger, ForeignKey("game.id"), nullable = False)
   game = relationship("Game")
   #rappresenta il dealer del round, ovvero chi sceglie le categorie!
-  dealer_id = Column(Integer, ForeignKey("user.id"), nullable = False)
+  dealer_id = Column(BigInteger, ForeignKey("user.id"), nullable = False)
   dealer = relationship("User")
   #categoria scelta per il round
-  cat_id = Column(Integer, ForeignKey("category.id"))
+  cat_id = Column(BigInteger, ForeignKey("category.id"))
   chosen_category = relationship("Category")
 
   __table_args__ = (UniqueConstraint('game_id', 'number', name = "unique_game_number"), )
@@ -62,10 +62,10 @@ class Quiz(Base, CommonPK):
   #la risposta giusta, Vero/Falso
   answer = Column(Boolean, nullable = False)
   #l'immagine di riferimento, se presente
-  image_id = Column(Integer, ForeignKey("image.id"))
+  image_id = Column(BigInteger, ForeignKey("image.id"))
   image = relationship("Image")
 	#la categoria di riferimento (si punta sul fatto che ogni quiz sia categorizzato)
-  category_id = Column(Integer, ForeignKey("category.id"), nullable = False)
+  category_id = Column(BigInteger, ForeignKey("category.id"), nullable = False)
   category = relationship("Category")
 
   export_properties = ["answered_correctly", "my_answer"]
@@ -74,24 +74,24 @@ class Quiz(Base, CommonPK):
 
 class Question(Base):
   #quiz estratto
-  quiz_id = Column(Integer, ForeignKey("quiz.id"), nullable = False, primary_key = True)
+  quiz_id = Column(BigInteger, ForeignKey("quiz.id"), nullable = False, primary_key = True)
   quiz = relationship("Quiz")
   #risposta data dall'utente
   answer = Column(Boolean, nullable = False)
 
-  round_id = Column(Integer, ForeignKey("round.id"), nullable = False, primary_key = True)
-  user_id = Column(Integer, ForeignKey("user.id"), nullable = False, primary_key = True)
+  round_id = Column(BigInteger, ForeignKey("round.id"), nullable = False, primary_key = True)
+  user_id = Column(BigInteger, ForeignKey("user.id"), nullable = False, primary_key = True)
   export_properties = ["round_number"]
 
 class Invite(Base):
     #il game a cui sei stato invitato
-    game_id = Column(Integer, ForeignKey("game.id"), nullable = False, primary_key = True)
+    game_id = Column(BigInteger, ForeignKey("game.id"), nullable = False, primary_key = True)
     game = relationship("Game")
     #chi ha invitato
-    sender_id = Column(Integer, ForeignKey("user.id"), nullable = False, primary_key = True)
+    sender_id = Column(BigInteger, ForeignKey("user.id"), nullable = False, primary_key = True)
     sender = relationship("User", foreign_keys = [sender_id])
     #chi ha ricevuto l'invito
-    receiver_id = Column(Integer, ForeignKey("user.id"), nullable = False, primary_key = True)
+    receiver_id = Column(BigInteger, ForeignKey("user.id"), nullable = False, primary_key = True)
     receiver = relationship("User", foreign_keys = [receiver_id])
     #l'invito è stato accettato? NB: questo valore ragiona a logica ternaria (True|False|NULL), in quanto un invito potrebbe non essere ne stato accettato ne rifiutato, e quindi il valore diventa NULL
     accepted = Column(Boolean, nullable = True, default = None)
@@ -120,13 +120,13 @@ class Category(Base, CommonPK):
 #categoria proposta al giocatore, che deve sceglierla
 class ProposedCategory(Base):
   #categoria proposta
-  category_id = Column(Integer, ForeignKey("category.id"), nullable = False, primary_key = True)
+  category_id = Column(BigInteger, ForeignKey("category.id"), nullable = False, primary_key = True)
   #turno in cui è stata proposta
-  round_id = Column(Integer, ForeignKey("round.id"), nullable = False, primary_key = True)
+  round_id = Column(BigInteger, ForeignKey("round.id"), nullable = False, primary_key = True)
 
 #domande proposte al giocatore per un turno specifico
 class ProposedQuestion(Base):
   #categoria proposta
-  quiz_id = Column(Integer, ForeignKey("quiz.id"), nullable = False, primary_key = True)
+  quiz_id = Column(BigInteger, ForeignKey("quiz.id"), nullable = False, primary_key = True)
   #turno in cui è stata proposta
-  round_id = Column(Integer, ForeignKey("round.id"), nullable = False, primary_key = True)
+  round_id = Column(BigInteger, ForeignKey("round.id"), nullable = False, primary_key = True)
