+++ tp/preferences/models.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from tp.base.models import *
-from sqlalchemy import Column, Integer, ForeignKey, Boolean, Enum
+from sqlalchemy import Column, Integer, ForeignKey, Boolean, Enum, BigInteger
 from sqlalchemy.orm import relationship
 
 class Preferences(Base, CommonPK):
@@ -15,5 +15,5 @@ class Preferences(Base, CommonPK):
     notification_message = Column(Boolean, nullable = False, default = True)
     notification_full_hearts = Column(Boolean, nullable = False, default = True)
     #user
-    user_id = Column(Integer, ForeignKey("user.id"), nullable = False)
+    user_id = Column(BigInteger, ForeignKey("user.id"), nullable = False)
     user = relationship("User")
